# Optimized metric name extraction in distributor