--- a/pkg/util/extract/extract.go
+++ b/pkg/util/extract/extract.go
@@ -14,12 +14,24 @@ var (
 )
 
 // MetricNameFromLabelAdapters extracts the metric name from a list of LabelPairs.
+// The returned metric name string is a copy of the label value.
 func MetricNameFromLabelAdapters(labels []cortexpb.LabelAdapter) (string, error) {
+	unsafeMetricName, err := UnsafeMetricNameFromLabelAdapters(labels)
+	if err != nil {
+		return "", err
+	}
+
+	// Force a string copy since LabelAdapter is often a pointer into
+	// a large gRPC buffer which we don't want to keep alive on the heap.
+	return string([]byte(unsafeMetricName)), nil
+}
+
+// UnsafeMetricNameFromLabelAdapters extracts the metric name from a list of LabelPairs.
+// The returned metric name string is a reference to the label value (no copy).
+func UnsafeMetricNameFromLabelAdapters(labels []cortexpb.LabelAdapter) (string, error) {
 	for _, label := range labels {
 		if label.Name == model.MetricNameLabel {
-			// Force a string copy since LabelAdapter is often a pointer into
-			// a large gRPC buffer which we don't want to keep alive on the heap.
-			return string([]byte(label.Value)), nil
+			return label.Value, nil
 		}
 	}
 	return "", errNoMetricNameLabel