@pracucci 2021-03-24
> Is there more performance improvements to squeeze out?

There's always room for improvement, but I will stop here for now :)